add omegaconf, replace {instance, toot_listen_msg} with it, refactor out model_load
