--- config_example/access_token.txt
@@ -1,1 +1,0 @@
-YOUR_ACCESS_TOKEN_GOES_HERE
