--- config_example/endpoint_url.txt
@@ -1,1 +1,0 @@
-https://YOUR_DOMAIN.COM.ES_HERE
