+++ config_example/instance.yaml
@@ -1,0 +1,2 @@
+endpoint_url: https://YOUR_DOMAIN.COM.ES_HERE
+access_token: YOUR_ACCESS_TOKEN_GOES_HERE
