+++ config_example/toot_listen_msg.yaml
@@ -1,0 +1,4 @@
+toot_listen_start: "#bot #bot_message\ndiffusers_mastodon_bot starts!"
+toot_listen_start_cw: "bot start"
+# End message may not be fired on exit.
+toot_listen_end: "#bot #bot_message\ndiffusers_mastodon_bot ends!"
