+++ diffusers_mastodon_bot/conf/instance_conf.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class InstanceConf:
+    endpoint_url: str
+    access_token: str
