+++ diffusers_mastodon_bot/main.py
@@ -6,87 +6,23 @@ import json
 from diffusers.utils.import_utils import is_xformers_available
 
 from mastodon import Mastodon
-
-import torch
-
-from diffusers.pipelines.stable_diffusion import StableDiffusionPipeline, StableDiffusionImg2ImgPipeline
+from omegaconf import OmegaConf
 
 from diffusers_mastodon_bot.app_stream_listener import AppStreamListener
 from diffusers_mastodon_bot.bot_request_handlers.bot_request_handler import BotRequestHandler
 from diffusers_mastodon_bot.bot_request_handlers.game.diffuse_game_handler import DiffuseGameHandler
 from diffusers_mastodon_bot.bot_request_handlers.diffuse_me_handler import DiffuseMeHandler
 from diffusers_mastodon_bot.bot_request_handlers.diffuse_it_handler import DiffuseItHandler
-from diffusers_mastodon_bot.community_pipeline.lpw_stable_diffusion \
-    import StableDiffusionLongPromptWeightingPipeline as StableDiffusionLpw
+from diffusers_mastodon_bot.conf.conf_helper import load_structured_conf_yaml
+
+
+from diffusers_mastodon_bot.conf.instance_conf import InstanceConf
+from diffusers_mastodon_bot.conf.toot_listen_msg_conf import TootListenMsgConf
+
+from diffusers_mastodon_bot.model_load import create_diffusers_pipeline
+
 
 logger = logging.getLogger(__name__)
-
-def create_diffusers_pipeline(device_name='cuda', pipe_kwargs: Optional[Dict[str, Any]] = None):
-    if pipe_kwargs is None:
-        pipe_kwargs = {}
-
-    pipe_kwargs = pipe_kwargs.copy()
-
-    kwargs_defaults = {
-        "pretrained_model_name_or_path": 'hakurei/waifu-diffusion',
-        'revision': 'fp16'
-    }
-
-    for key, value in kwargs_defaults.items():
-        if key not in pipe_kwargs:
-            pipe_kwargs[key] = value
-
-    model_name_or_path = pipe_kwargs['pretrained_model_name_or_path']
-    del pipe_kwargs['pretrained_model_name_or_path']
-
-    torch_dtype = torch.float32
-    if 'torch_dtype' in pipe_kwargs:
-        dtype_param = pipe_kwargs['torch_dtype']
-        del pipe_kwargs['torch_dtype']
-
-        if dtype_param == 'torch.float16':
-            torch_dtype = torch.float16
-
-    if 'scheduler' in pipe_kwargs:
-        scheduler_param = pipe_kwargs['scheduler']
-        del pipe_kwargs['scheduler']
-
-        if scheduler_param == 'euler':
-            from diffusers import EulerDiscreteScheduler
-            pipe_kwargs['scheduler'] = EulerDiscreteScheduler.from_pretrained(model_name_or_path, subfolder="scheduler")
-        elif scheduler_param == 'euler_a':
-            from diffusers import EulerAncestralDiscreteScheduler
-            pipe_kwargs['scheduler'] = EulerAncestralDiscreteScheduler.from_pretrained(model_name_or_path,
-                                                                                       subfolder="scheduler")
-        elif scheduler_param == 'dpm_solver++':
-            from diffusers import DPMSolverMultistepScheduler
-            pipe_kwargs['scheduler'] = DPMSolverMultistepScheduler.from_pretrained(model_name_or_path,
-                                                                                   subfolder="scheduler")
-
-    pipe: StableDiffusionLpw = StableDiffusionLpw.from_pretrained(
-        model_name_or_path,
-        torch_dtype=torch_dtype,
-        safety_checker=None,
-        **pipe_kwargs
-    )
-
-    pipe = pipe.to(device_name)
-    pipe.enable_attention_slicing()
-
-    if is_xformers_available():
-        try:
-            pipe.unet.enable_xformers_memory_efficient_attention(True)
-        except Exception as e:
-            logger.warning(
-                "Could not enable memory efficient attention. Make sure xformers is installed"
-                f" correctly and a GPU is available: {e}"
-            )
-
-    pipe_kwargs['pretrained_model_name_or_path'] = model_name_or_path
-    pipe_kwargs['torch_dtype'] = 'torch.float16' if torch_dtype == torch.float16 else 'torch.float32'
-    pipe_kwargs['scheduler'] = str(type(pipe.scheduler).__name__)
-
-    return pipe, pipe_kwargs
 
 
 def read_text_file(filename: str) -> Union[str, None]:
@@ -119,20 +55,8 @@ def main():
         ]
     )
 
-    access_token = read_text_file('./config/access_token.txt')
-    endpoint_url = read_text_file('./config/endpoint_url.txt')
-
-    if access_token is None:
-        print('mastodon access token is required but not found. check ./config/access_token.txt')
-        exit()
-
-    if access_token is None:
-        print('mastodon endpoint url is required but not found. check ./config/endpoint_url.txt')
-        exit()
-
-    toot_listen_start = read_text_file('./config/toot_listen_start.txt')
-    toot_listen_end = read_text_file('./config/toot_listen_end.txt')
-    toot_listen_start_cw = read_text_file('./config/toot_listen_start_cw.txt')
+    instance: InstanceConf = load_structured_conf_yaml(InstanceConf, './config/instance.yaml')  # type: ignore
+    toot_listen_msg: TootListenMsgConf = load_structured_conf_yaml(TootListenMsgConf, './config/toot_listen_msg.yaml')  # type: ignore
 
     pipe_kwargs = load_json_dict('./config/pipe_kwargs.json')
     proc_kwargs = load_json_dict('./config/proc_kwargs.json')
@@ -144,8 +68,8 @@ def main():
 
     logger.info('starting')
     mastodon = Mastodon(
-        access_token=access_token,
-        api_base_url=endpoint_url
+        access_token=instance.access_token,
+        api_base_url=instance.endpoint_url
     )
 
     logger.info('info checking')
@@ -179,17 +103,17 @@ def main():
     ]  # type: ignore
 
     logger.info('creating listener')
-    listener = AppStreamListener(mastodon, pipe,
-                                 mention_to_url=my_url,
-                                 req_handlers=req_handlers,
-                                 toot_listen_start=toot_listen_start,
-                                 toot_listen_start_cw=toot_listen_start_cw,
-                                 toot_listen_end=toot_listen_end,
-                                 device=device_name,
-                                 proc_kwargs=proc_kwargs,
-                                 pipe_kwargs=pipe_kwargs,
-                                 **app_stream_listener_kwargs
-                                 )
+    listener = AppStreamListener(
+        mastodon_client=mastodon,
+        diffusers_pipeline=pipe,
+        mention_to_url=my_url,
+        req_handlers=req_handlers,
+        toot_listen_msg=toot_listen_msg,
+        device=device_name,
+        proc_kwargs=proc_kwargs,
+        pipe_kwargs=pipe_kwargs,
+        **app_stream_listener_kwargs
+    )
 
     mastodon.stream_user(listener, run_async=False, timeout=10000)
